services/max-parser/requirements.txt
maxapi-python>=1.2.4
motor>=3.3.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pymongo>=4.6.0
